// src/types/formTypes.tsx
export interface FormFieldOption {
  value: string;
  label: string;
  dataTestId?: string;
}

export interface Validation {
  message: string;
}

export interface FormField {
  fieldId: string;
  type:
    | "text"
    | "tel"
    | "email"
    | "textarea"
    | "date"
    | "dropdown"
    | "radio"
    | "checkbox";
  label: string;
  placeholder?: string;
  required: boolean;
  dataTestId: string;
  validation?: Validation;
  options?: FormFieldOption[];
  maxLength?: number;
  minLength?: number;
}

export interface FormSection {
  sectionId: number;
  title: string;
  description: string;
  fields: FormField[];
}

export interface FormResponse {
  message: string;
  form: {
    formTitle: string;
    formId: string;
    version: string;
    sections: FormSection[];
  };
}